+++ src/main/java/com/laytonsmith/abstraction/events/MCEntityDamageEvent.java
@@ -14,7 +14,7 @@ public interface MCEntityDamageEvent extends BindableEvent {
 
     public MCEntity getEntity();
 
-    public double getDamage();
+    public int getDamage();
 
-    public void setDamage(double damage);
+    public void setDamage(int damage);
 }
